Fix local build
This patch fixes local build of gcpproxy from the following error:
In file included from ./src/envoy/http/service_control/config_parser.h:21:0,
                 from ./src/envoy/http/service_control/filter.h:21,
                 from src/envoy/http/service_control/filter_factory.cc:17:
./src/api_proxy/path_matcher/path_matcher.h:181:27: error: 'google::api_proxy::path_matcher::PathMatcherNode::PathInfo google::api_proxy::path_matcher::{anonymous}::TransformHttpTemplate(const google::api_proxy::path_matcher::HttpTemplate&)' defined but not used [-Werror=unused-function]
 PathMatcherNode::PathInfo TransformHttpTemplate(const HttpTemplate& ht) {
                           ^~~~~~~~~~~~~~~~~~~~~
cc1plus: all warnings being treated as errors

Obviously we use this function but it looks like a bug in the compiler failing
to recognize that it is used in Register(). Adding a simple namespace fixes it.
Given that we are able to apply a simple fix, I am not super interested in
reporting this bug to the bazel team.

Bug: n/a
Test: prow/gcpproxy-build.sh
Change-Id: Ie48f4c8d51b9e74e57cae24b304b02d962cf3814