Add bash commands to generate data-plane-api go protos and their dependencies
We should not use gogo-generated go protos in our repo.
This change introduces a script to automatically generate go protos with the official goproto library via a bazel target.
It also copies over dependencies to the correct paths for the protos to work as expected.
Minor changes to the Makefile and test scripts in order to prevent automatic deletion of generated files.

Change-Id: Ic67834cc91394b1a36e44d5a15398b8dfc782daf