Modify imports to use goproto generated envoy *.pb.go
Modified imports to reference generated go protos.
Added conventional naming to proto imports (It is conventional in google3 to rename protos to "pb" or some other short abbreviation ending in "pb").
Minor changes to source code (pointers, dereferences, goproto-to-goto conversions).

This is a large change. It is impossible to break this into smaller changes due to the incompatibility between goproto and gogo.
Note that this change introduces runtime conversion from goproto to gogo. Implications of this include:
- Runtime inefficiency
- Possibility of undefined behavior at runtime

A note about directory structure:
- `go-control-plane` and `gogo/protobuf` should ONLY be referenced by files in `go/src/configmanager/`. No other files should use these packages, as they rely on `gogo`
- `data-plane-api` and `golang/protobuf` can be referenced from any go files

Testing Done:
- Successful copybara to `google3/third_party`, where all unit tests pass (after manually replicating upstream envoy changes)

Change-Id: Ie01b3436e91365d7a7a64e81ac16ae29a843855b