Update go to version 1.13
Blocking the port of configmanager into google3 because it depends on go module support.

Also fix broken test framework:
"Testing flags are now registered in the new Init function, which is invoked by the generated main function for the test. As a result, testing flags are now only registered when running a test binary, and packages that call flag.Parse during package initialization may cause tests to fail."

Change-Id: I5241b0f85a1f0a83de199988cdf4fdaffdcf3b42