b/138815331: Format logs to integrate with Stackdriver Logging
Output logs in glog format. This allows users to filter by timestamp and severity, and makes logs a lot nicer to read.

https://screenshot.googleplex.com/EuQK6OMYSsN

This change will fix 90% of the formatting issues with logs. I still need to make a change to upstream envoy to address the formatting issues with the other 10% of log lines.

Ref: https://github.com/envoyproxy/envoy/issues/8637

Change-Id: Ic50c4ea3a0a7b289318b4785c93d2fab6c258ffe