b/143774323: change prow image to have clang
This image has clang-8 and clang-8 will be set in CC and CXX. The
only problem is we cannot use clang with -c opt and bazel cache same
time as b/143672285. So, we will still use gcc for CI/CD and release
but use clang for sanitizers.

This change should be merged after prow image change for presubmit job.

Change-Id: Ic1e39fa32961e2dbddb5af020d89423c244c2758