b/148231316: Fix brittle tests broken by change in Go 1.14
The test targets fail when built with Go 1.14 due to reliance on the exact text of "net/url" errors. Go 1.14 quotes the text of the URL in errors to improve clarity when the URL contains spaces.

Tests should not rely on the exact text of errors: http://go/go-test-comments#test-error-semantics

Unsure if any of our integration tests will break. But let's get this fixed first to unblock Google3 usage.

Signed-off-by: Teju Nareddy <nareddyt@google.com>
Change-Id: I70bf3dc01da306d34f4220c7b311a412cdb50147