b/148106246: add cloud-run-on-anthos e2e test
The test job will create a gke cluster, deploy cloud run on it and run
the test cases. It test the case that cloudesf run on cloud run of anthos
using http.

To be note: the normal cases like api-key, jwt-token are all covered but
the stress test doesn't work currently(b/148950591).

Change-Id: I33c1b189d4c9bdcb210d8e42e9c84af99deaa176