b/147834348: Auto-generate JWT Audience if no audiences are set in service config
If no audience is set, generate the audience `https://$SERVICE_NAME`. This follows the publicly documented behavior of ESPv1.

A few JWT integration tests relied on audiences not being checked. New tokens with the correct audiences are now used for those tests.

New integration tests for this default value are located at the end of `TestGRPCJwt`.

Change-Id: I98e20dfd3ef47aca2decb65369daa7ff1ce96320
Signed-off-by: Teju Nareddy <nareddyt@google.com>