b/149050012: Remove backend protocol flag (#31)
Allow the user to configure the entire backend address (scheme, host, and port) in the --backend flag. Remove the need for the --backend_protocol flag.

This maintains backwards compatibility with ESPv2 via the default backend address.

Actual code changes are very minimal. Majority of changes are to unit tests and integration tests:

BackendProtocol was required in almost all of the config generator package.
Every single integration test sets the --backend_protocol flag. Remove the flag in all integration tests. Just set the flag when starting config manager.
Also fix service_control_failed_request_report_test.go. In that test, backend protocol was configured incorrectly. It was using a GRPC backend but configuring the protocol as HTTP.

Signed-off-by: Teju Nareddy nareddyt@google.com