Handle null pointer for running envoy in a named pipe (#71)
`Address` can either be a IP address or a named pipe. We try to read from a `nullptr` if the address is a named pipe. Even though we will never run into this problem with the current deployment model of ESPv2, we should still handle this case for other open-source users and simpler fuzzing.

Added reproducer test-case.

Ref: https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=21174
Signed-off-by: Teju Nareddy <nareddyt@google.com>