Switch /src/api_proxy to envoy_build_system
This allows coverage charts for unit tests located in these directories.

Unfortunately, I was unable to switch over /src/api_proxy/auth:json_util_test because it directly depends on grpc_internals that do not compile with `-Wall`. This should be OK, as this is a small part of our library.

Signed-off-by: Teju Nareddy <nareddyt@google.com>