Fix network fail open for Unavailable CheckResponse errors (#152)
If SC returns unavailable as part of the `CheckResponse` proto (not the HTTP call), then we always network fail open. Modify the logic so we only network fail open when the option is true. While we don't expect many customers to disable network fail open, it gives us a chance to log warnings and fixes a bug.

Also refactor the callback a little. This should make it easier to track stats for check responses (#151, b/155209585, etc.).

Testing: Add integration test for `network_fail_open` with 5xx vs non-5xx `CheckResponse` errors.

Signed-off-by: Teju Nareddy <nareddyt@google.com>