b/155209585 #1: Refactor callback for testability and implement `denied_control_plane_fault` (#155)
When HTTP call is unsuccessful, SC Check returns HTTP 500, or SC Check returns a check response with a 5xx-like status:
- For network fail open: Increment `allowed_control_plane_fault`
- For network fail closed: Increment `refactor_control_plane_fault`

The code changes for above are very simple (2 lines to increment a counter). Most of this diff is from splitting the callback into another function that we can unit test. Added unit tests for blocked/allowed requests and the stats.

Signed-off-by: Teju Nareddy <nareddyt@google.com>