b/155209585 #5: Verify stat invariants for all integration tests
Stats recording in SC Filter is split across multiple files and can be confusing to reason about. To ensure that requests are not double-counted in SLO-related stats, add a `stats_verifier` component in our integration tests. This component will fetch envoy stats at the end of each test and verify no invariants are broken.

Example invariants:
> The values of the detailed `service_control.denied_*` counters must sum up to `service_control.denied`

We can add more invariants as needed in the future.

Signed-off-by: Teju Nareddy <nareddyt@google.com>