b/158262661: Move snake to json segment mapping to per-operation instead of global (#218)
In OpenAPI specs, a path template jsonName may conflict with the jsonName of well known `google.protobuf.*` types. To prevent this, define the snake_to_json segment mapping per operation instead of globally in the path matcher filter. Then the validation of name conflicts will only be per-operation, greatly making the validation less strict.

Also optimize when this needs to be configured in the path matcher config. We only pass this when we need path parameter extraction under constant address backend routing.

Diff is too large already. I will follow-up with PRs for:
- Adding integration tests.
- Incrementing api version (breaking API change).

Ref: https://groups.google.com/u/1/g/google-cloud-endpoints/c/hxyuDx1yrvQ
Closes #185

Signed-off-by: Teju Nareddy <nareddyt@google.com>