Switch e2e tests to test tracing with W3C trace context propagation (#252)
GPC's serverless offerings will all start supporting `traceparent` headers soon for integration with OpenTelemetry. go/w3c-trace-context-guidelines

Switch our e2e tests to also use this header for trace context propagation. This will help give a better view of networking latency mentioned in #237. For each compute platform, we can determine exactly how much time is spent between the request leaving ESPv2 and entering the backend.

Here is an example of a cold start that we can now detect because of this PR. My plan is to let this run in our e2e tests for a while, and the evaluate if we should make this the default option for `--tracing_incoming_context` and `--tracing_outgoing_context`. It will make it easier for API Producers to debug cold starts.

![image](https://user-images.githubusercontent.com/11142171/92157065-e604ca00-edf7-11ea-8b2e-cc51df4ffbf8.png)

Signed-off-by: Teju Nareddy <nareddyt@google.com>