Add decoration to the route names for dynamic routing (#319)
This results in better naming for traces and more details in `--enable_debug`.

Currently we only add it for routes for dynamic routing. This is the higher priority, since we care about traces for serverless platforms due to the network hop. Once we move over to use Envoy's path matcher for all our routes, we can support it for sidecar routes too.

Testing done:
- Manually verified trace names are correct for our e2e tests.
- Integration tests.

Signed-off-by: Teju Nareddy <nareddyt@google.com>