Envoy route config for sidecar with x-google-backend (#354)
Configuration for ESPv2 in sidecar mode, but with `x-google-backend` defined to configure timeouts per route.

NOTE: Currently we **do not generate the correct config**, the `deadline` in `x-google-backend` is ignored in sidecar mode. But let's check in this config for now, it will make it easier for me to reason about what needs to change in config generator. And it will give me a good test case for per-route config implementation.

Ref #321 (does not fix the issue)