b/174609312 #1: Handle trailing backslash for regex-match paths
Currently, ESPv2 only matches paths when they are not followed by a trailing slash. This matches the path template syntax, but is problematic for some clients. Transition ESPv2 to match paths with trailing slashes. This PR only changes regex-based paths - it is trivial and will minimize the diff of the following PR.

The trailing slash will **not** be stripped from the request unless path rewrite is configured. We need to be careful with `APPEND_PATH_TO_ADDRESS` path rewrites to prevent duplicate slashes.

Signed-off-by: Teju Nareddy <nareddyt@google.com>