b/175352046: Enable fallback to x-cloud-trace-context by default (#439)
Cloud Run currently uses `X-Cloud-Trace-Context` for context propagation. If we enable incoming trace context to fallback to this header (when `traceparent` is missing), the request path will be tied to App servers traces. We also enable it as an outgoing trace context so App servers backends are also captured.

Example of App Server traces for both the ESPv2 and Backend instance:
![image](https://user-images.githubusercontent.com/11142171/101963218-300e0080-3bdc-11eb-8e29-6769eae28a77.png)

Note: If the backend uses `traceparent` headers already, then Cloud Trace may not correctly correlate the AppServer traces with the backend ones. This will result in disjoint traces. We don't need to worry about this, no data will be lost.

Implementation detail:
- We move setting the defaults from `start_proxy.py` to config manager. This ensures Cloud API Gateway will receive the same defaults.

Testing done:
- Updated integration test to capture this behavior

Signed-off-by: Teju Nareddy <nareddyt@google.com>