Switch integration tests to use 127.0.0.1 instead of localhost (#485)
When running in google3, we need to be explicit about using IPv4 vs IPv6. `localhost` will choose IPv6 by default, but ESPv2 is meant to run on GCP with IPv4 addresses. So we modify our integration test framework to explicitly use IPv4 addresses wherever possible.

Diff is large, but there is no substantial change. It is just replacing all hard-coded "localhost" with a call to a common function. Done manually...