refactor: add new `RouteGenerator` for CORS and Health Check (#845)
Similar to `FilterGenerator` and `ClusterGenerator`, add the concept of a `RouteGenerator`.

Start with a few trivial generators and helpers. Much of the code is copied from `service_info.go` and `route_generator.go`.

No production change, these files are tested but not used yet. Copied code will be deleted in follow-up.

Signed-off-by: Teju Nareddy <nareddyt@google.com>