refactor: implement RouteGenerator for backend routes (#848)
`BackendGenerator` is a `RouteGenerator` to configure routes to the local or remote backend service. It configures all aspects of the routes that are forward to a backend cluster, including sorting them via HTTP pattern trie for correctness. Please see `backend_test.go` to double check the generated routes.

Fork a few methods from `ServiceInfo` into utility file. Refactor methods to return `map[selector]X` as pre-processed input for the generator.

TODO in follow-up CL:
- CORS pass-through routes to backends (when `AllowCors` is enabled).
- Set route deadlines and timeouts.

NOT used in prod yet

Signed-off-by: Teju Nareddy <nareddyt@google.com>